Move the fibonacci calculator to random projects
