+++ CSharp/RandomProjects/RandomProjects.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32319.34
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Numbers", "Numbers\Numbers.csproj", "{C325117B-3594-4139-B0FA-5D4E23972FC8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Numbers.Specs", "Numbers.Specs\Numbers.Specs.csproj", "{056761F2-FBB9-4EF7-8EB4-40D46EB64FEE}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C325117B-3594-4139-B0FA-5D4E23972FC8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C325117B-3594-4139-B0FA-5D4E23972FC8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C325117B-3594-4139-B0FA-5D4E23972FC8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C325117B-3594-4139-B0FA-5D4E23972FC8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{056761F2-FBB9-4EF7-8EB4-40D46EB64FEE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{056761F2-FBB9-4EF7-8EB4-40D46EB64FEE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{056761F2-FBB9-4EF7-8EB4-40D46EB64FEE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{056761F2-FBB9-4EF7-8EB4-40D46EB64FEE}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A3DA16E3-5BA6-4B92-9CC9-EF4454A3A77F}
+	EndGlobalSection
+EndGlobal
